Add script to clean mesa
This script is based on commands given to me by dllud and trial
and error to cleanup some additional generated objects.

This script has been made because the clean-% targets are not
sufficient to completely cleanup mesa:
- make clean-libmesa_nir doesn't result in a rebuild of the nir
  generated files.
- make clean-libGLES_mesa is only one of the many LOCAL_MODULE
  targets and is not sufficient either to clean it all up.

So while we might still miss some part in this script, it cleans
at least more than a single clean target.

As we progress, we might want to replace some of the rm -rf in
that script by make clean-<target> if we are confident enough
that they do the same job or a better one.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>