Add preliminary build wrapper script
Replicant 6.0 is now build with a script that wraps the build process.
If we don't have any blockers like build failure, this makes it easier
for new developers to build Replicant. It is also used by the release
scripts.

It also brings several advantages along the way:
- It enables us to simplify the build instructions while still making
  it simple enough to get the lower level commands by reading the
  script.
- It saves build logs by default.
- It enables us to add additional commands, for instance to workaround
  a build issue, without needing to update all the build instructions.
- We can build all targets in one row.

As we Replicant 10 is still a work in progress, we still need to add
support for signing images in Replicant 10 and in this script.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>