clean-mesa.sh: unify interface with build.sh
Users already used to build.sh will find that interface more natural and
less error prone as the probability of making mistakes with the machine
target is lower since the interface is the same.

A sanity checks with rm -rf was also added along the way, and that command
was moved at the end for more safety as it's easier to spot if it's taking
too long. It also gives more time to react after, in the event where a bug
makes that command do damage.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>