Kernel build task: Use the host xz
The prebuilt xz doesn't have the --arm option:
    $ ./prebuilts/build-tools/linux-x86/bin/xz --arm --lzma2=,dict=32M
      ./prebuilts/build-tools/linux-x86/bin/xz: Unsupported filter chain or filter options

While the host has it:
    $ xz --arm --lzma2=,dict=32M
      xz: Compressed data cannot be written to a terminal
      xz: Try `xz --help' for more information.

The --arm option is needed for building the kernel as it is used in
the scripts/xz_wrap.sh script from the kernel sources.

However we cannot simply whitelist the host xz and remove the
prebuilt xz to force its use by the kernel build system as it will
create issues like that with other packages:
    [ 35% 81/227] including external/mesa3d/Android.mk ...
    external/mesa3d/src/mesa/Android.libmesa_glsl_utils.mk:74:
    [...]/prebuilts/build-tools/linux-x86/bin/ckati was modified,
	regenerating...
    FAILED: ninja: 'prebuilts/build-tools/linux-x86/bin/xz', needed by
	'out/soong/.intermediates/bionic/linker/ld-android/
	 android_arm_armv7-a-neon_cortex-a9_shared/ld-android.so',
	 missing and no known rule to make it.

We also need to respect the host's PATH instead of forcing XZ to
/bin/xz for instance as the location of the binary changes
depending on the distribution used. For instance in Guix it will
not be in /bin/xz but rather somewhere in /gnu.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>